Renaming sub-app folder to shopify-app
